Department and Users Module completed

+++ src/app/modules/user/user-view/user-view.component.ts
@@ -1,13 +1,16 @@
 import { Component, OnInit, OnDestroy, AfterViewInit, ViewChild } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { MatPaginator, MatTableDataSource } from '@angular/material';
+import { MatPaginator, MatTableDataSource, MatSnackBar } from '@angular/material';
 import { Subscription } from 'rxjs';
-import { Router } from '@angular/router';
+import { Router, ActivatedRoute } from '@angular/router';
 
 import { DataSharingService } from '../../../public service/data-sharing.service';
 import { ToastrService } from 'ngx-toastr';
 import { UserService } from '../service/user.service';
 import { NgxSpinnerService } from 'ngx-spinner';
+import { DialogService } from '../../../public service/dialog.service';
+import { SpinnerService } from '../../../public service/spinner.service';
+import { AppDialogData } from '../../../model/appDialogData';
 
 @Component({
   selector: 'app-user-view',
@@ -23,24 +26,34 @@ export class UserViewComponent implements AfterViewInit, OnDestroy {
   totalCount = 0;
   userData: any = {};
   departmentId: any;
+  Router: any;
+  deptId: any;
+  isAlreadySubscribedToDialogUserActionService: boolean = false;
+
+
 
   displayedColumns: string[] = ['userId', 'userName', 'profileImage', 'departmentTitle', 'mobileNumber', 'emailId', 'userRoleIdFK', 'Actions'];
-  paidDataSource: MatTableDataSource<User> = new MatTableDataSource(); 
+  paidDataSource: MatTableDataSource<User> = new MatTableDataSource();
 
   //@ViewChild('paidPaginator') paidPaginator: MatPaginator;
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
   previousSubscription: Subscription;
   upcomingSubscription: Subscription;
-  Router: any;
+  message: any = {};
+  DepartmentObj: string = '';
+  userId: number;
+
 
   constructor(private http: HttpClient,
     private userService: UserService,
     private router: Router,
+    private route: ActivatedRoute,
     public dataService: DataSharingService,
-    private toastr: ToastrService,
-    private spinner:NgxSpinnerService  
-    
+    private dialogService: DialogService,
+    private snackBar: MatSnackBar,
+    private spinnerService: SpinnerService,
+
   ) {
 
   }
@@ -52,27 +65,56 @@ export class UserViewComponent implements AfterViewInit, OnDestroy {
   }
 
   ngOnInit() {
-    this.getAllUsers(this.pageNumber); 
+    //this.deptId = this.route.snapshot.params['departmentId'];
+    if (this.message && this.message.departmentId) {
+      console.log('event init', this.message);
+      this.DepartmentObj = '(' + this.message.name + ')';
+      this.getAllUsers(this.message.departmentId, this.pageNumber);
+    }
+  }
+
+  receiveMessage($event) {
+    this.message = $event;
+    console.log('event emited', this.message);
+    if (this.message.departmentTitle) {
+      this.DepartmentObj = '(' + this.message.departmentTitle + ')';
+    }
+    if (this.message.name) {
+      this.DepartmentObj = '(' + this.message.name + ')';
+    }
+    this.getAllUsers(this.message.departmentId, this.pageNumber);
+  }
+
+
+
+  isNoRecord() {
+    return this.totalCount == 0;
   }
 
   ngOnDestroy(): void { }
 
   /*********************************************************** Get All Users *******************************************************************/
 
-  getAllUsers(pageNo: any) {
-    this.spinner.show();
-    setTimeout(() => {
-      this.spinner.hide();
-    }, 1000);
-    this.userService.getAllUsers(pageNo).subscribe(res => { 
-      console.log(res);
-      this.paidDataSource=res.users;
-      this.pageNumber=res.currentPage;
-      this.totalCount=res.totalCount; 
+  getAllUsers(departmentId: number, pageNo: number) {
+    this.spinnerService.setSpinnerVisibility(true);
+    this.userService.getAllUsers(departmentId, pageNo).subscribe(res => {
+      this.spinnerService.setSpinnerVisibility(false);
+      if (res.users) {
+        console.log(res);
+        this.paidDataSource = res.users;
+        this.pageNumber = res.currentPage;
+        this.totalCount = res.totalCount;
+      } else {
+        this.showSnackBar(res.message);
+      }
     },
-    error =>{
-      this.toastr.error(error.error.message);
-    })
+      error => {
+        this.showSnackBar("Something went wrong..!!");
+      })
+  }
+
+  showSnackBar(message: string) {
+    this.snackBar.open(message, '', { duration: 2000 });
   }
 
 
@@ -82,28 +124,48 @@ export class UserViewComponent implements AfterViewInit, OnDestroy {
   pageChange(pageNo: any) {
     this.loading = true;
     this.page = pageNo.pageIndex;
-    this.getAllUsers(this.page);
+    this.getAllUsers(this.message.departmentId, this.page);
   }
 
   /*********************************************************** Edit Particular Asset  *******************************************************************/
 
-  editUser(userId: number) {
-    this.dataService.changeData(userId);
+  editUser(visit) {
+    this.dataService.changeData(visit);
     this.router.navigate(['/user/add-user']);
   }
 
 
 
   /*********************************************************** Delete Particular Asset *******************************************************************/
-  deleteUser(userId: number) {
-    alert('are you sure?');
-    this.userService.deleteUser(userId).subscribe(res => {
-      this.toastr.success(res.message);
-      this.getAllUsers(this.page);
-    })
-    error => {
-      console.log(error);
-      this.toastr.error(error.message);
+
+  deleteUser(userId: number, firstName: string, lastName: string) {
+    this.userId = userId;
+    let appDialogData: AppDialogData = {
+      visibilityStatus: true,
+      title: 'DELETE USER',
+      message: ` Are your sure you want to delete user "${firstName} ${lastName}"`,
+      positiveBtnLable: "Yes",
+      negativeBtnLable: "Cancel"
+    }
+    this.dialogService.setDialogVisibility(appDialogData);
+    if (!this.isAlreadySubscribedToDialogUserActionService) {
+      this.isAlreadySubscribedToDialogUserActionService = true;
+      this.dialogService.getUserDialogAction().subscribe(userAction => {
+        if (userAction == 0) {
+          //User has not performed any action on opened app dialog or closed the dialog;
+        } else if (userAction == 1) {
+          this.dialogService.setUserDialogAction(0);
+          //User has approved delete operation
+          this.spinnerService.setSpinnerVisibility(true);
+          this.userService.deleteUser(userId).subscribe(res => {
+            this.spinnerService.setSpinnerVisibility(false);
+            this.showSnackBar(res.message);
+            this.getAllUsers(this.message.departmentId, this.page);
+          }, error => {
+            this.showSnackBar("Something went wrong..!!");
+          });
+        }
+      })
     }
   }
 
@@ -112,7 +174,7 @@ export class UserViewComponent implements AfterViewInit, OnDestroy {
   addUser() {
     let selectedUser = null;
     this.dataService.changeData(selectedUser);
-    this.router.navigate(['/user/add-user']); 
+    this.router.navigate(['/user/add-user']);
   }
 
 
@@ -122,9 +184,10 @@ export class UserViewComponent implements AfterViewInit, OnDestroy {
 
 export interface User {
   userId: number,
-  userName: string,
+  firstName: string,
+  lastName: string,
   profileImage: string,
-  userRoleIdFK:number,
+  userRoleIdFK: number,
   departmentTitle: string,
   mobileNumber: string,
   emailId: string,
